fix bug for repositories
TO DO : vérfiier nullité d’un produit demandé

+++ src/main/java/com/rizomm/misys/controller/ProductController.java
@@ -21,7 +21,7 @@ public class ProductController implements ErrorController {
     private ProductRepository _productRepository;
 
     @RequestMapping(value="/detail/{id}", method = RequestMethod.GET)
-    public ModelAndView detailProduct(@PathVariable Long id){
+    public ModelAndView detailProduct(@PathVariable int id){
         Product product = _productRepository.findOne(id);
         return new ModelAndView("product/detail", "product", product);
     }
